Added child classes in different moduls
